+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   avoid weird effects from unstable framerate. This means that the proper
   numbers for it should be greatly reduced -  for example in the 3D example it
   was reduced from 60.0 to 1.2.
+- Better document how to shapecast and why it is needed. Also make it the
+  default for the examples.
 
 
 ## 0.1.0 - 2023-03-16